Move prisma.ts into the prisma folder
